+++ src/utils/addStepsToDemo.ts
@@ -1,22 +1,29 @@
-import { Icons, Step } from "../models";
+import { DemoFile, Icons, Step } from "../models";
 import { DemoPanel } from "../panels/DemoPanel";
 import { DemoCreator, FileProvider } from "../services";
 
-export const addStepsToDemo = async (steps: Step | Step[], title?: string, description?: string, icons?: Icons) => {
-  const demoFile = await FileProvider.demoQuickPick();
-  if (!demoFile?.demo) {
+export const addStepsToDemo = async (
+  steps: Step | Step[],
+  demoFile?: {
+    filePath: string;
+    demo: DemoFile;
+  },
+  title?: string,
+  description?: string,
+  icons?: Icons
+) => {
+  if (!demoFile?.filePath || !demoFile?.demo) {
     return;
   }
-  const { filePath, demo } = demoFile;
 
-  const updatedDemos = await DemoCreator.askWhereToAddStep(demo, steps, title, description, icons);
-  if (!updatedDemos) {
+  const demoFileToUpdate = await DemoCreator.askWhereToAddStep(demoFile.demo, steps, title, description, icons);
+  if (!demoFileToUpdate) {
     return;
   }
 
-  demo.demos = updatedDemos;
+  demoFile.demo.demos = demoFileToUpdate.demos;
 
-  await FileProvider.saveFile(filePath, JSON.stringify(demo, null, 2));
+  await FileProvider.saveFile(demoFile.filePath, JSON.stringify(demoFile.demo, null, 2));
 
   // Trigger a refresh of the treeview
   DemoPanel.update();