+++ src/utils/chooseDemoFile.ts
@@ -1,0 +1,14 @@
+import { FileProvider } from "../services";
+
+export const chooseDemoFile = async () => {
+  const demoFile = await FileProvider.demoQuickPick();
+  if (!demoFile?.demo) {
+    return;
+  }
+  const { filePath, demo } = demoFile;
+
+  return {
+    filePath,
+    demo,
+  };
+};