+++ src/utils/getFileContents.ts
@@ -1,12 +1,18 @@
 import { Uri, WorkspaceFolder, workspace } from "vscode";
 import { General } from "../constants";
+import { DemoRunner } from "../services";
 
 export const getFileContents = async (workspaceFolder: WorkspaceFolder, contentPath?: string) => {
   if (!contentPath) {
     return;
   }
 
-  const contentUri = Uri.joinPath(workspaceFolder.uri, General.demoFolder, contentPath);
+  const version = DemoRunner.getCurrentVersion();
+
+  const contentUri =
+    version === 2
+      ? Uri.joinPath(workspaceFolder.uri, contentPath)
+      : Uri.joinPath(workspaceFolder.uri, General.demoFolder, contentPath);
   if (!contentUri) {
     return;
   }